Datepickers reusable hook for time value
